Consider: pkg_resources / setuptools is great to generate RPM but py2exe sucks with setuptools. pkg_resources can't give a path for a resource in a zip file (obvious). py2exe doesn't like the __file__ var (seems obvious too). Wapiti must be able to run from a RPM install. It must also be able to run from a decompressed tarball without knowing about it (so it must find where it is). A py2exe package of Wapiti would be great.
So:
I kept setuptools for the setup.py, having good RPMs is nice.
I removed the pkg_resource dependencies in the Wapiti code, now I just
import a module from wapitiCore then get the path of this module.
I check for the "frozen" state to know if Wapiti is running from a
generated py2exe package.
If not frozen I still check on __file__ for decompressed tarball.
I added a make_exe.py script for py2exe (looks good so far but users
will have to force 0 to --verify-ssl because the CERT file is not
included)

Plus:
I removed the .py extension to the 2 cookie utilities.
Fixed some platform related bug in Wapiti (not really an error but
wasn't nice to see)
Fixed some english translations and added german ones (thanks int23h !)



git-svn-id: http://svn.code.sf.net/p/wapiti/code/trunk@312 6f1da90a-7915-0410-b0c8-fe3d2c9d8697